Refacto: separate the 'pendings' queue in client.js in a 'pendings' queue and a 'waitings' queue
The pendings queue stores the calls that are waiting for a server to
connect to our client.
The waitings queue stores calls that have been already made to the
server and are awaiting a response.

This made for a better clarity of the code, and avoid possible race
condition if a disconnected/connected event happens during a wait for a
response.