Fix #2 - Implement a server killing mechanism
Smuggler is now able to unregister a server and kill a worker/iframe.

Basically, server sends an unregister request when they have no
connected client left after a disconnection.

Then, the smuggler checks if this server is in use for another contract.
If not, it terminates/kills the Worker or the iframe.

Note that neither SharedWorker nor ServiceWorkers are killed. They are
merely unregistered.