WIP Fixes #3 - Kill/restart logic server on visibilitychange
This successfully removes the iframe when the app goes to background and
add it back when the app goes to foreground.

However, it does not reestablish a functional application. My best guess
so far : the killed iframe containing the server declared some clients
this server was using. These clients needs to disconnect before we can
cleanly kill this iframe.