Reworked app-goes-to-background process
Instead of unloading/reloading only the 'logic' server we now kill all servers.
This implicitely disconnect all clients as well.

On restore, we only need to take care of one client (window.logicAPI). The others
will be recreated on the fly (either on js reload or when their servers are built).