Fix the reload of clients when the server goes back to live.
Before, disconnecting a client involved removing listener to the
'server' and deleting reference to the 'server'. Actually, this 'server'
is nothing more than a BroadcastChannel to listen to server event.
Therefore we should not delete it, as we receive 'connected' events by
this mean.

This commit keeps the reference to this broadcast channel, and renames
this variable to better reflect what it really is.