Make client.connect and client.disconnect return a promise
As a consequence, we now have a finer control about the connection and
disconnection process: we can wait for these process to complete, we know
if a connection or a disconnection is processing and can deal with
connection or disconnection requests accordingly.

One should remember that we can have 'connected' and 'disconnected'
events without requesting it though. At the moment, if we receive a
connection event, we act as if we are connected whether we requested it
or not. This is debatable though.