Fixes #16 - Fix onclick listeners to work across reload of the app
We just reference the method by its name instead of using a pointer to
it. This pointer are not a reliable way to reference these methods,
because the prototype get changed across unload/reload of the
application (prototype gets deleted on unload, and recreated on reload.
This allows us to go into 'record' mode when there is no server
available).